+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.2] - 2023-02-21
+
+### Added
+
+- Added Luxembourgish (lb) translation (#2); thank you Alain Fontaine
+- Added French (fr) translation (#3); thank you Alain Fontaine
+- Added Finnish (fi) translation (PR#3); thank you Thomas Raehalme
+
 ## [1.1.1] - 2023-01-23
 
 ### Added